timers are now char

+++ chip8_emulator/src/lib/OpCodes/OpCodeFX07.java
@@ -10,6 +10,6 @@ import lib.MachineState;
 public class OpCodeFX07 extends OpCodeVX {
     @Override
     protected void execute(MachineState mc, int myCode, int vX) {
-        mc.V[vX] = mc.delay_timer;
+        mc.V[vX] = (byte) mc.delay_timer;
     }
 }
